requirement.txt
fastapi
uvicorn[standard]
python-multipart
passlib
python-jose[cryptography]
motor
bcrypt
websockets
motor